refactor(backend): extract tenancy mixins and shared serializer fields into farm/common
Move ProjectScopedMixin/ProjectRevisionMixin from farm/views.py into
farm/common/mixins.py and AuditUserSerializer/CentimetersField/
LocalizedDecimalField/_resolve_active_project_from_serializer from
farm/serializers.py into farm/common/serializer_fields.py. Verbatim
moves with import repoints only; no behavior changes.
