+++ backend/farm/common/__init__.py
@@ -1,0 +1,1 @@
+"""Shared API plumbing used by all farm domain packages."""
