+++ backend/farm/serializers.py
@@ -6,6 +6,12 @@ from django.core.exceptions import ValidationError
 from django.db import IntegrityError, transaction
 from rest_framework import serializers
 
+from .common.serializer_fields import (
+    AuditUserSerializer,
+    CentimetersField,
+    LocalizedDecimalField,
+    _resolve_active_project_from_serializer,
+)
 from .enum_normalization import normalize_seed_rate_unit
 from .seed_units import (
     SEED_PACKAGE_UNIT_GRAMS,
@@ -44,19 +50,6 @@ BED_NAME_DUPLICATE_MESSAGE = 'Ein Beet mit diesem Namen existiert in dieser Parz
 EMPTY_SEED_RATE_UNIT_VALUES = {None, '', '-'}
 
 
-def _resolve_active_project_from_serializer(serializer) -> Project | None:
-    """Resolve active project from serializer context or bound instance."""
-    request = serializer.context.get('request')
-    if request is not None:
-        active_project = getattr(request, 'active_project', None)
-        if active_project is not None:
-            return active_project
-    instance = getattr(serializer, 'instance', None)
-    if instance is not None and hasattr(instance, 'project'):
-        return instance.project
-    return None
-
-
 def _normalize_seed_rate_unit_value(value: object) -> str | None:
     """Normalize supported seed rate units and legacy empty placeholders."""
     if value in EMPTY_SEED_RATE_UNIT_VALUES:
@@ -65,57 +58,6 @@ def _normalize_seed_rate_unit_value(value: object) -> str | None:
     if normalized_value:
         return normalized_value
     raise serializers.ValidationError('Unsupported seed rate unit.')
-
-
-class AuditUserSerializer(serializers.Serializer):
-    id = serializers.IntegerField()
-    email = serializers.EmailField()
-    display_name = serializers.SerializerMethodField()
-    display_label = serializers.SerializerMethodField()
-
-    def get_display_name(self, obj) -> str:
-        full_name = f'{obj.first_name or ""} {obj.last_name or ""}'.strip()
-        return full_name or obj.username
-
-    def get_display_label(self, obj) -> str:
-        full_name = self.get_display_name(obj)
-        if full_name:
-            return f'{full_name} ({obj.email})'
-        return obj.email or obj.username
-
-
-class CentimetersField(serializers.FloatField):
-    """Expose meter-based model fields as centimeters in the API."""
-    
-    def to_representation(self, value):
-        if value is None:
-            return None
-        return float(value) * 100.0
-    
-    def to_internal_value(self, data):
-        cm_value = super().to_internal_value(data)
-        return cm_value / 100.0
-
-
-class LocalizedDecimalField(serializers.DecimalField):
-    """Decimal field that accepts comma decimals and returns float JSON values."""
-
-    default_error_messages = {
-        'invalid': 'Please enter a valid numeric value, e.g. 3.9.',
-    }
-
-    def to_internal_value(self, data):
-        normalized = data
-        if isinstance(data, str):
-            normalized = data.strip().replace(',', '.')
-        return super().to_internal_value(normalized)
-
-    def to_representation(self, value):
-        decimal_value = super().to_representation(value)
-        if decimal_value is None:
-            return None
-        return float(decimal_value)
-
 
 
 class LocationSerializer(serializers.ModelSerializer):
